Generalize `get_image_matrix` function
- Allow not to store pixel data in the polars
dataframe resulting from `get_image_matrix`.
This is useful for deep learning models, where
PyTorch DataSet class is better for efficiently
handling images in 3D tensors. This is done
through `get_sobel` argument, which if true will
result in a polars dataframe with image paths.

- Minor fix in `set_up_date` method of `DataExtractor`.
This generalizes the random seed to the one
selected by the user.